% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicle_scale.R
\name{logicle_trans}
\alias{logicle_trans}
\title{Logicle transformation for ggplot2.
See ?flowCore::logicletGml2 for the definition of the parameters.}
\usage{
logicle_trans(w = 0.5, t = 262144, m = 4.5)
}
\description{
Logicle transformation for ggplot2.
See ?flowCore::logicletGml2 for the definition of the parameters.
}

